requirements.txt
requests
jinja2
PyYAML
beautifulsoup4
Pillow
